requirements.txt
python-binance==1.0.19
xgboost>=2.0.0
ta>=0.11.0
pandas>=2.0.0
numpy>=1.24.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
httpx>=0.25.0
optuna>=3.0.0
scikit-learn>=1.3.0
pyyaml>=6.0
